add --no-delete

+++ tasks/index_migrate.go
@@ -19,6 +19,7 @@ const (
 type IndexMigrateOptions struct {
 	ESClient         *elastic.Client
 	COSClient        *cos.Client
+	NoDelete         bool
 	Dir              string
 	Index            string
 	Bulk             int
@@ -30,7 +31,7 @@ func (opts IndexMigrateOptions) Workspace() string {
 	return filepath.Join(opts.Dir, opts.Index)
 }
 
-func IndexMigratie(opts IndexMigrateOptions) conc.Task {
+func IndexMigrate(opts IndexMigrateOptions) conc.Task {
 	return conc.TaskFunc(func(ctx context.Context) (err error) {
 		log.Printf("确保工作目录: %s", opts.Workspace())
 		if err = os.MkdirAll(opts.Workspace(), 0755); err != nil {
@@ -61,9 +62,11 @@ func IndexMigratie(opts IndexMigrateOptions) conc.Task {
 		if err = conc.ParallelWithLimit(opts.Concurrency, tasks...).Do(ctx); err != nil {
 			return
 		}
-		log.Printf("删除索引: %s", opts.Index)
-		if _, err = opts.ESClient.DeleteIndex(opts.Index).Do(ctx); err != nil {
-			return
+		if !opts.NoDelete {
+			log.Printf("删除索引: %s", opts.Index)
+			if _, err = opts.ESClient.DeleteIndex(opts.Index).Do(ctx); err != nil {
+				return
+			}
 		}
 		log.Printf("删除本地目录: %s", opts.Workspace())
 		if err = os.RemoveAll(opts.Workspace()); err != nil {
